refactor(protocol_designer): move Slideout from atoms to components
move Slideout from atoms to components

close AUTH-627
