+++ protocol-designer/src/components/Slideout/index.tsx
